+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [UNRELEASED]
+* [#426] Updated proposal tally result, added activeVotingPower value to active proposal collection.
 * [#524] Added `noreferrer` to Banner links
 
 ## v0.41.x-14